requirements.txt
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.1
python-dotenv==1.1.1
redis==6.4.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
